+++ e2e-test/cypress/integration/groups_spec.js
@@ -7,36 +7,36 @@ describe("User/Groups Dashboard", function() {
   it("Show group table", function() {
     cy.location("pathname").should("eq", "/users");
     cy.get("[aria-label=groupsTab]").click();
-    cy.get("#userdashboard").should("be.visible");
+    cy.get("[data-test=userdashboard]").should("be.visible");
   });
 
   it("Create new group", function() {
     cy.get("[data-test=create]").click();
-    cy.get("#id")
+    cy.get("[data-test=groupid] input")
       .type("TestGroup")
       .should("have.value", "TestGroup");
-    cy.get("#name")
+    cy.get("[data-test=groupname] input")
       .type("testgroup")
       .should("have.value", "testgroup");
-    cy.get("#autoComplete-input")
+    cy.get("[data-test=autocomplete] input")
       .type("mstein")
       .should("have.value", "mstein");
-    cy.get("#autoComplete-item-0")
+    cy.get("[data-test=autocomplete-item-0]")
       .should("be.visible")
       .should("have.text", "mstein")
       .click();
-    cy.get("#autoComplete-input")
+    cy.get("[data-test=autocomplete] input")
       .type("thouse")
       .should("have.value", "thouse");
-    cy.get("#autoComplete-item-0")
+    cy.get("[data-test=autocomplete-item-0]")
       .should("be.visible")
       .should("have.text", "thouse")
       .click();
-    cy.get("[aria-label=submit]").click();
+    cy.get("[data-test=submit]").click();
   });
 
   it("Created group should be visible", function() {
-    cy.get("#group-TestGroup")
+    cy.get("[data-test=group-TestGroup]")
       .find("th")
       .then($th => {
         expect($th).to.have.length(1);