+++ e2e-test/cypress/integration/users_spec.js
@@ -1,30 +1,99 @@
+const testUser = {
+  userName: "dviolin",
+  password: "test"
+};
+
 describe("Users/Groups Dashboard", function() {
   before(() => {
     cy.login();
     cy.visit("/users");
   });
 
-  it("Show user dashboard", function() {
+  it("User dashboard is displayed for user with basic permissions", function() {
+    cy.login(testUser.userName, testUser.password);
+    cy.visit("/users");
+
+    // User dashboard should be visible to this user
     cy.location("pathname").should("eq", "/users");
-    cy.get("#userdashboard").should("be.visible");
+    cy.get("[data-test=userdashboard]").should("be.visible");
+  });
+
+  it("Button to add user not visible to user without proper permission", function() {
+    // Login as user without permissions
+    cy.login(testUser.userName, testUser.password);
+    cy.visit("/users");
+
+    // User dashboard should be visible to this user
+    cy.location("pathname").should("eq", "/users");
+    cy.get("[data-test=userdashboard]").should("be.visible");
+
+    cy.get("[data-test=create]").should("not.exist");
+  });
+
+  it("Button to change password should be visible for all users", function() {
+    // Login as user with basic permissions
+    cy.login(testUser.userName, testUser.password);
+    cy.visit("/users");
+
+    // User dashboard should be visible to this user
+    cy.location("pathname").should("eq", "/users");
+    cy.get("[data-test=userdashboard]").should("be.visible");
+
+    cy.get(`[data-test=edit-user-${testUser.userName}]`).should("be.visible");
+  });
+
+  it("Button to change user permissions is not visible to user without proper permission", function() {
+    // Login as user without permissions
+    cy.login(testUser.userName, testUser.password);
+    cy.visit("/users");
+
+    // User dashboard should be visible to this user
+    cy.location("pathname").should("eq", "/users");
+    cy.get("[data-test=userdashboard]").should("be.visible");
+
+    cy.get("[data-test*=edit-permissions-]").should("not.exist");
+  });
+
+  it("Button to add user is visible to user with proper permission", function() {
+    // Login as user with permissions
+    cy.login();
+    cy.visit("/users");
+
+    // User dashboard should be visible to this user
+    cy.location("pathname").should("eq", "/users");
+    cy.get("[data-test=userdashboard]").should("be.visible");
+
+    cy.get("[data-test=create]").should("be.visible");
+  });
+
+  it("Button to edit user permissions is visible to user with proper permission", function() {
+    // Login as user with permissions
+    cy.login();
+    cy.visit("/users");
+
+    // User dashboard should be visible to this user
+    cy.location("pathname").should("eq", "/users");
+    cy.get("[data-test=userdashboard]").should("be.visible");
+
+    cy.get("[data-test*=edit-permissions-]").should("be.visible");
   });
 
   it("Create new user", function() {
     cy.get("[data-test=create]").click();
-    cy.get("#fullname")
+    cy.get("[data-test=fullname] input")
       .type("Test User")
       .should("have.value", "Test User");
-    cy.get("#username")
+    cy.get("[data-test=username] input")
       .type("testuser")
       .should("have.value", "testuser");
-    cy.get("#password")
+    cy.get("[data-test=password] input")
       .type("test")
       .should("have.value", "test");
-    cy.get("[aria-label=submit]").click();
+    cy.get("[data-test=submit]").click();
   });
 
   it("Created user should be visible", function() {
-    cy.get("#user-testuser")
+    cy.get("[data-test=user-testuser]")
       .find("th")
       .then($th => {
         expect($th).to.have.length(1);