+++ frontend/src/languages/french.js
@@ -188,7 +188,15 @@ const fr = {
     new_group: "Nouveau groupe d'utilisateurs",
     add_group: "Ajouter un groupe",
     edit_group: "Éditer und groupe",
-    edit_permissions_for: "Éditer les autorisations pour"
+    edit_permissions_for: "Éditer les autorisations pour",
+    current_user_password: "Votre mot de passe",
+    new_user_password: "Nouveau mot de passe",
+    new_user_password_confirmation: "Confirmer le nouveau mot de passe",
+    no_password_match: "Les mots de passe ne correspondent pas",
+    change_password_for: "Changer le mot de passe pour {0}",
+    password_change_success: "Mot de passe changé avec succès",
+    type_current_password: "Tapez le mot de passe actuel pour {0}",
+    type_new_password: "Tapez le nouveau mot de passe pour {0}"
   },
 
   nodesDashboard: {