+++ frontend/src/languages/german.js
@@ -186,7 +186,15 @@ const de = {
     group_created: "German: Group successfully created",
     add_group: "German: Add Group",
     edit_group: "German: Edit Group",
-    edit_permissions_for: "German: Edit permissions for"
+    edit_permissions_for: "German: Edit permissions for",
+    current_user_password: "Ihr Passwort",
+    new_user_password: "Neues Passwort",
+    new_user_password_confirmation: "Bestätigen Sie das neue Password",
+    no_password_match: "Passwörter stimmen nicht überein",
+    change_password_for: "Passwortänderung für {0}",
+    password_change_success: "Passwort wurde erfolgreich geändert",
+    type_current_password: "Geben Sie das Passwort für {0} ein",
+    type_new_password: "Geben Sie das neue Passwort für {0} ein"
   },
 
   nodesDashboard: {