+++ frontend/src/pages/Navbar/SideNav.js
@@ -3,15 +3,14 @@ import React from "react";
 import Drawer from "@material-ui/core/Drawer";
 
 import SideNavCard from "./SideNavCard";
-import { canViewUserDashboard, canViewNodesDashboard } from "../../permissions";
+import { canViewNodesDashboard } from "../../permissions";
 
 const SideNav = props => {
   const { showSidebar, toggleSidebar, allowedIntents, ...rest } = props;
-  const userDashboardEnabled = canViewUserDashboard(allowedIntents);
   const nodeDashboardEnabled = canViewNodesDashboard(allowedIntents);
   return (
     <Drawer anchor="left" open={showSidebar} onClose={toggleSidebar}>
-      <SideNavCard nodeDashboardEnabled={nodeDashboardEnabled} userDashboardEnabled={userDashboardEnabled} {...rest} />
+      <SideNavCard nodeDashboardEnabled={nodeDashboardEnabled} {...rest} />
     </Drawer>
   );
 };