+++ frontend/src/pages/Users/UsersTable.js
@@ -1,15 +1,17 @@
-import React from "react";
+import IconButton from "@material-ui/core/IconButton";
+import Paper from "@material-ui/core/Paper";
+import { withStyles } from "@material-ui/core/styles";
 import Table from "@material-ui/core/Table";
+import TableBody from "@material-ui/core/TableBody";
+import TableCell from "@material-ui/core/TableCell";
 import TableHead from "@material-ui/core/TableHead";
 import TableRow from "@material-ui/core/TableRow";
-import TableCell from "@material-ui/core/TableCell";
-import TableBody from "@material-ui/core/TableBody";
-import Paper from "@material-ui/core/Paper";
-import _sortBy from "lodash/sortBy";
+import EditIcon from "@material-ui/icons/Edit";
 import PermissionIcon from "@material-ui/icons/LockOpen";
-import IconButton from "@material-ui/core/IconButton";
+import _sortBy from "lodash/sortBy";
+import React from "react";
+
 import strings from "../../localizeStrings";
-import { withStyles } from "@material-ui/core/styles";
 
 const styles = {
   iconColor: {
@@ -20,8 +22,8 @@ const sortUsers = users => {
   return _sortBy(users, user => user.organization && user.id);
 };
 
-const UsersTable = ({ classes, users, permissionIconDisplayed, showDashboardDialog, userId, allowedIntents }) => {
-  const sortedUsers = sortUsers(users.filter(u => u.isGroup !== true)); //&& u.id !== userId));
+const UsersTable = ({ classes, users, permissionIconDisplayed, showDashboardDialog, userId }) => {
+  const sortedUsers = sortUsers(users.filter(u => u.isGroup !== true));
 
   return (
     <Paper>
@@ -36,8 +38,15 @@ const UsersTable = ({ classes, users, permissionIconDisplayed, showDashboardDial
         </TableHead>
         <TableBody id="usertablebody">
           {sortedUsers.map(user => {
+            const canEditPassword =
+              // need to check if user permissions exist yet
+              // to make sure this is compatible with older versions
+              user.permissions &&
+              user.permissions.hasOwnProperty("user.changePassword") &&
+              user.permissions["user.changePassword"].some(x => x === userId);
+
             return (
-              <TableRow id={`user-${user.id}`} key={user.id}>
+              <TableRow data-test={`user-${user.id}`} key={user.id}>
                 <TableCell component="th" scope="row">
                   {user.id}
                 </TableCell>
@@ -45,8 +54,19 @@ const UsersTable = ({ classes, users, permissionIconDisplayed, showDashboardDial
                 <TableCell>{user.organization}</TableCell>
                 <TableCell>
                   {permissionIconDisplayed ? (
-                    <IconButton onClick={() => showDashboardDialog("editUser", user.id)}>
+                    <IconButton
+                      data-test={`edit-permissions-${user.id}`}
+                      onClick={() => showDashboardDialog("editUserPermissions", user.id)}
+                    >
                       <PermissionIcon className={classes.iconColor} />
+                    </IconButton>
+                  ) : null}
+                  {canEditPassword ? (
+                    <IconButton
+                      data-test={`edit-user-${user.id}`}
+                      onClick={() => showDashboardDialog("editUserPassword", user.id)}
+                    >
+                      <EditIcon />
                     </IconButton>
                   ) : null}
                 </TableCell>