+++ CHANGELOG.md
@@ -9,15 +9,33 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 <!-- ### Added-->
 
-<!-- ### Changed-->
+<!-- ### Changed -->
 
 <!-- ### Deprecated -->
 
 <!-- ### Removed -->
 
-<!-- ### Fixed-->
+<!-- ### Fixed -->
 
 <!-- ### Security -->
+
+## [1.2.0] - 2019-06-26
+
+### Added
+
+- UI feature to change a user's password [#325](https://github.com/openkfw/TruBudget/issues/325)
+
+### Changed
+
+- Every user can see the users and groups table [#333](https://github.com/openkfw/TruBudget/issues/333)
+
+### Removed
+
+- Removed permissions related to notifications [#316](https://github.com/openkfw/TruBudget/issues/316)
+
+### Fixed
+
+- Display correct currency when editing workflow items [#281](https://github.com/openkfw/TruBudget/issues/281)
 
 ## [1.1.0] - 2019-06-12
 
@@ -80,8 +98,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `/project.viewHistory` deprecated in favor of `/project.viewHistory.v2`. [#252](https://github.com/openkfw/TruBudget/issues/252)
 - `/subproject.viewHistory` deprecated in favor of `/subproject.viewHistory.v2`. [#252](https://github.com/openkfw/TruBudget/issues/252)
 
-<!-- ### Removed -->
-
 ### Fixed
 
 - Fixed line of YAML file for master deployments via docker-compose, so that image of excel export service is pulled correctly [#223](https://github.com/openkfw/TruBudget/issues/223)
@@ -89,8 +105,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Budgets on project analytics do not contain open workflow items [#230](https://github.com/openkfw/TruBudget/issues/230)
 - Fixed a bug where on smaller screens the action buttons (create & cancel) are hidden and no item could be created [#240](https://github.com/openkfw/TruBudget/issues/240)
 - Increase frontend stability [#263](https://github.com/openkfw/TruBudget/pull/263)
-
-<!-- ### Security -->
 
 ## [1.0.0-beta.9] - 2019-04-23
 
@@ -297,7 +311,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Updated translation keys and language-specific formatting.
 - Fixed bug where the subproject permissions dialog would break the details view of another project.
 
-[unreleased]: https://github.com/openkfw/TruBudget/compare/v1.1.0...master
+[unreleased]: https://github.com/openkfw/TruBudget/compare/v1.2.0...master
+[1.2.0]: https://github.com/openkfw/TruBudget/compare/v1.1.0...v1.2.0
 [1.1.0]: https://github.com/openkfw/TruBudget/compare/v1.0.1...v1.1.0
 [1.0.1]: https://github.com/openkfw/TruBudget/compare/v1.0.0...v1.0.1
 [1.0.0]: https://github.com/openkfw/TruBudget/compare/v1.0.0-beta.9...v1.0.0